Extract DamageType Enum from Server
