+++ megamek/src/megamek/common/weapons/ACFlechetteHandler.java
@@ -23,8 +23,8 @@ import megamek.common.TargetRoll;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.common.options.OptionsConstants;
+import megamek.server.DamageType;
 import megamek.server.Server;
-import megamek.server.Server.DamageType;
 
 /**
  * @author Andrew Hunter
