+++ megamek/src/megamek/common/weapons/MGHandler.java
@@ -26,7 +26,7 @@ import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.common.options.OptionsConstants;
 import megamek.server.Server;
-import megamek.server.Server.DamageType;
+import megamek.server.DamageType;
 
 /**
  * @author Andrew Hunter
