+++ megamek/src/megamek/common/weapons/SRMAXHandler.java
@@ -23,7 +23,7 @@ import megamek.common.Game;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.server.Server;
-import megamek.server.Server.DamageType;
+import megamek.server.DamageType;
 
 /**
  * @author Sebastian Brocks
