+++ megamek/src/megamek/common/weapons/SRMAntiTSMHandler.java
@@ -19,7 +19,7 @@ import megamek.common.*;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.common.options.OptionsConstants;
 import megamek.server.Server;
-import megamek.server.Server.DamageType;
+import megamek.server.DamageType;
 
 /**
  * @author Sebastian Brocks
