+++ megamek/src/megamek/common/weapons/battlearmor/BAMGHandler.java
@@ -24,7 +24,7 @@ import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.common.weapons.WeaponHandler;
 import megamek.server.Server;
-import megamek.server.Server.DamageType;
+import megamek.server.DamageType;
 
 /**
  * @author Sebastian Brockxs
