+++ megamek/src/megamek/server/DamageType.java
@@ -1,0 +1,6 @@
+package megamek.server;
+
+public enum DamageType {
+    NONE, FRAGMENTATION, FLECHETTE, ACID, INCENDIARY, INFERNO, IGNORE_PASSENGER,
+    ANTI_TSM, ANTI_INFANTRY, NAIL_RIVET, NONPENETRATING
+}
